validator: check exist user for delete

+++ src/main/java/com/example/employee_assessment_system/service/UserService.java
@@ -74,6 +74,7 @@ public class UserService {
         return userMapper.toDTO(userRepository.save(user));
     }
     public ApiResponse deleteUser(Long id) {
+        if (!userRepository.findById(id).isPresent()) throw new EntityNotFoundException("Not found user");
         userRepository.deleteById(id);
         return new ApiResponse(false,"Deleted user successfully",null);
     }
